backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.27.2
python-jose==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
